delete all data that is associated to a particular entity;
e.g. if a profile is deleted, all categories, tasks and subtasks and questions associated will be deleted. if a category is removed the all tasks and subtasks and questions associated will be removed. if a task is removed than all associated subtasks will also be removed. This, there's no lost data associated to nothing in the database. This also solves the issue where if something is deleted and the remains are left behind, then the next entity created will have the same id as the previous, and all associated data will show up on the new entity;